cibfortune/requirements.txt
torch>=2.0.0
transformers>=4.57.0
accelerate
sentencepiece
protobuf
Pillow
requests
gradio>=4.0.0
